Restore reporting routes module
