+++ routes/__init__.py
@@ -1,619 +1,42 @@
-from datetime import date
+"""Top-level API router for the application.
 
-from fastapi import APIRouter, Request, Form, UploadFile, File
-from fastapi.responses import HTMLResponse, RedirectResponse
+This package used to contain all route handlers in a single module.  To make
+the codebase easier to navigate the handlers have been split into multiple
+modules grouped by area of responsibility (authentication, inventory and
+reporting).  Each module exposes its own ``APIRouter`` instance which is
+included in the router defined here.
+"""
 
-from utils import templates, load_settings, save_settings, get_table_columns
-from models import (
-    SessionLocal,
-    StockItem,
-    PrinterInventory,
-    LicenseInventory,
-    HardwareInventory,
-    RequestItem,
-    LookupItem,
-    User,
-    pwd_context,
-)
+from fastapi import APIRouter, Request
+from fastapi.responses import RedirectResponse
 
 router = APIRouter()
 
 
 def require_login(request: Request) -> RedirectResponse | None:
-    """Redirect to the login page when the user is unauthenticated."""
+    """Redirect unauthenticated users to the login page.
+
+    Many endpoints require an authenticated session.  They call this helper and
+    redirect the user when the session does not contain a ``"user"`` key.
+    """
+
     if not request.session.get("user"):
         return RedirectResponse("/login", status_code=303)
     return None
 
-# Mapping from short table identifiers used in the UI to the actual
-# SQLAlchemy model classes. This allows API endpoints to operate on
-# different tables in a generic manner.
-MODEL_MAP = {
-    "stock": StockItem,
-    "printer": PrinterInventory,
-    "license": LicenseInventory,
-    "inventory": HardwareInventory,
-}
+
+# Import and register sub-routers.  The imports must happen after the helper
+# above is defined so the modules can import it without causing circular
+# imports.
+from .authentication import router as authentication_router  # noqa: E402
+from .inventory import router as inventory_router  # noqa: E402
+from .reporting import router as reporting_router  # noqa: E402
 
 
-@router.get("/", response_class=HTMLResponse)
-def root(request: Request) -> HTMLResponse:
-    """Render the main dashboard page.
+# Expose all API endpoints through a single router to the FastAPI application.
+router.include_router(authentication_router)
+router.include_router(inventory_router)
+router.include_router(reporting_router)
 
-    The template expects a number of context variables. Provide
-    sensible defaults so the page can render even when the database
-    is empty or not yet configured.
-    """
+__all__ = ["router", "require_login"]
 
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "factories": {},
-        "actions": [],
-        "type_labels": [],
-        "type_counts": [],
-    }
-    return templates.TemplateResponse(request, "main.html", context)
-
-
-@router.get("/login", response_class=HTMLResponse)
-def login_page(request: Request, error: str | None = None) -> HTMLResponse:
-    """Render the login page."""
-
-    return templates.TemplateResponse(request, "login.html", {"error": error})
-
-
-@router.post("/login")
-def login(request: Request, username: str = Form(...), password: str = Form(...)):
-    """Basic form-based login."""
-
-    db = SessionLocal()
-    try:
-        user = db.query(User).filter(User.username == username).first()
-        if user and pwd_context.verify(password, user.password):
-            request.session["user"] = user.username
-            return RedirectResponse("/", status_code=303)
-    finally:
-        db.close()
-    return templates.TemplateResponse(
-        request,
-        "login.html",
-        {"error": "Invalid credentials"},
-        status_code=401,
-    )
-
-
-@router.get("/logout")
-def logout(request: Request):
-    """Log the user out and redirect to the login page."""
-    request.session.clear()
-    return RedirectResponse("/login", status_code=303)
-
-
-@router.get("/stock", response_class=HTMLResponse)
-def stock_page(request: Request) -> HTMLResponse:
-    """Render the stock tracking page with empty defaults."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "stocks": [],
-        "columns": [],
-        "column_widths": {},
-        "lookups": {},
-        "offset": 0,
-        "page": 1,
-        "total_pages": 1,
-        "q": "",
-        "per_page": 25,
-        "table_name": "stock",
-        "filters": [],
-    }
-    return templates.TemplateResponse(request, "stok.html", context)
-
-
-@router.get("/stock/status", response_class=HTMLResponse)
-def stock_status_page(request: Request) -> HTMLResponse:
-    """Render simple stock status page."""
-    if redirect := require_login(request):
-        return redirect
-    return templates.TemplateResponse(request, "stok_durumu.html")
-
-
-@router.post("/stock/add")
-async def stock_add(request: Request):
-    """Create or update a stock item from form data."""
-    if redirect := require_login(request):
-        return redirect
-    form = await request.form()
-    db = SessionLocal()
-    try:
-        stock_id = form.get("stock_id")
-        if stock_id:
-            item = db.query(StockItem).get(int(stock_id))
-            if item:
-                for field in [
-                    "urun_adi",
-                    "kategori",
-                    "marka",
-                    "adet",
-                    "departman",
-                    "guncelleme_tarihi",
-                    "islem",
-                    "tarih",
-                    "ifs_no",
-                    "aciklama",
-                    "islem_yapan",
-                ]:
-                    if field in form:
-                        value = form.get(field)
-                        if field in {"adet"}:
-                            value = int(value) if value else None
-                        elif field in {"guncelleme_tarihi", "tarih"}:
-                            value = date.fromisoformat(value) if value else None
-                        setattr(item, field, value)
-        else:
-            item = StockItem(
-                urun_adi=form.get("urun_adi"),
-                kategori=form.get("kategori"),
-                marka=form.get("marka"),
-                adet=int(form.get("adet") or 0),
-                departman=form.get("departman"),
-                guncelleme_tarihi=
-                    date.fromisoformat(form.get("guncelleme_tarihi"))
-                    if form.get("guncelleme_tarihi")
-                    else None,
-                islem=form.get("islem"),
-                tarih=
-                    date.fromisoformat(form.get("tarih"))
-                    if form.get("tarih")
-                    else None,
-                ifs_no=form.get("ifs_no"),
-                aciklama=form.get("aciklama"),
-                islem_yapan=form.get("islem_yapan"),
-            )
-            db.add(item)
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/stock", status_code=303)
-
-
-@router.get("/printer", response_class=HTMLResponse)
-def printer_page(request: Request) -> HTMLResponse:
-    """Render the printer inventory page with empty defaults."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "printers": [],
-        "columns": [],
-        "column_widths": {},
-        "offset": 0,
-        "page": 1,
-        "total_pages": 1,
-        "q": "",
-        "per_page": 25,
-        "table_name": "printer",
-        "filters": [],
-    }
-    return templates.TemplateResponse(request, "yazici.html", context)
-
-
-@router.post("/printer/add")
-async def printer_add(request: Request):
-    """Create or update a printer inventory item."""
-    if redirect := require_login(request):
-        return redirect
-    form = await request.form()
-    db = SessionLocal()
-    try:
-        printer_id = form.get("printer_id")
-        if printer_id:
-            item = db.query(PrinterInventory).get(int(printer_id))
-            if item:
-                for field in [
-                    "yazici_markasi",
-                    "yazici_modeli",
-                    "kullanim_alani",
-                    "ip_adresi",
-                    "mac",
-                    "hostname",
-                    "tarih",
-                    "islem_yapan",
-                    "notlar",
-                ]:
-                    if field in form:
-                        value = form.get(field)
-                        if field == "tarih":
-                            value = date.fromisoformat(value) if value else None
-                        setattr(item, field, value)
-        else:
-            item = PrinterInventory(
-                yazici_markasi=form.get("yazici_markasi"),
-                yazici_modeli=form.get("yazici_modeli"),
-                kullanim_alani=form.get("kullanim_alani"),
-                ip_adresi=form.get("ip_adresi"),
-                mac=form.get("mac"),
-                hostname=form.get("hostname"),
-                tarih=
-                    date.fromisoformat(form.get("tarih"))
-                    if form.get("tarih")
-                    else None,
-                islem_yapan=form.get("islem_yapan"),
-                notlar=form.get("notlar"),
-            )
-            db.add(item)
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/printer", status_code=303)
-
-
-@router.post("/printer/upload")
-async def printer_upload(request: Request, excel_file: UploadFile = File(...)):
-    """Accept a printer inventory Excel upload (currently discarded)."""
-    if redirect := require_login(request):
-        return redirect
-    await excel_file.read()
-    return RedirectResponse("/printer", status_code=303)
-
-
-@router.get("/home", response_class=HTMLResponse)
-def home_page(request: Request) -> HTMLResponse:
-    """Render the dashboard from the /home path."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "factories": {},
-        "actions": [],
-        "type_labels": [],
-        "type_counts": [],
-    }
-    return templates.TemplateResponse(request, "main.html", context)
-
-
-@router.get("/inventory", response_class=HTMLResponse)
-def inventory_page(request: Request) -> HTMLResponse:
-    """Render the hardware inventory page."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "items": [],
-        "columns": [],
-        "column_widths": {},
-        "lookups": {},
-        "offset": 0,
-        "page": 1,
-        "total_pages": 1,
-        "q": "",
-        "per_page": 25,
-        "table_name": "inventory",
-        "filters": [],
-        "count": 0,
-    }
-    return templates.TemplateResponse(request, "envanter.html", context)
-
-
-@router.post("/inventory/add")
-async def inventory_add(request: Request):
-    """Create or update a hardware inventory record."""
-    if redirect := require_login(request):
-        return redirect
-    form = await request.form()
-    db = SessionLocal()
-    try:
-        item_id = form.get("item_id")
-        if item_id:
-            item = db.query(HardwareInventory).get(int(item_id))
-            if item:
-                for field in [
-                    "no",
-                    "fabrika",
-                    "blok",
-                    "departman",
-                    "donanim_tipi",
-                    "bilgisayar_adi",
-                    "marka",
-                    "model",
-                    "seri_no",
-                    "sorumlu_personel",
-                    "kullanim_alani",
-                    "bagli_makina_no",
-                    "ifs_no",
-                    "tarih",
-                    "islem_yapan",
-                ]:
-                    if field in form:
-                        value = form.get(field)
-                        if field == "tarih":
-                            value = date.fromisoformat(value) if value else None
-                        setattr(item, field, value)
-        else:
-            item = HardwareInventory(
-                no=form.get("no"),
-                fabrika=form.get("fabrika"),
-                blok=form.get("blok"),
-                departman=form.get("departman"),
-                donanim_tipi=form.get("donanim_tipi"),
-                bilgisayar_adi=form.get("bilgisayar_adi"),
-                marka=form.get("marka"),
-                model=form.get("model"),
-                seri_no=form.get("seri_no"),
-                sorumlu_personel=form.get("sorumlu_personel"),
-                kullanim_alani=form.get("kullanim_alani"),
-                bagli_makina_no=form.get("bagli_makina_no"),
-                ifs_no=form.get("ifs_no"),
-                tarih=
-                    date.fromisoformat(form.get("tarih"))
-                    if form.get("tarih")
-                    else None,
-                islem_yapan=form.get("islem_yapan"),
-            )
-            db.add(item)
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/inventory", status_code=303)
-
-
-@router.post("/inventory/upload")
-async def inventory_upload(request: Request, excel_file: UploadFile = File(...)):
-    """Accept a hardware inventory Excel upload (currently discarded)."""
-    if redirect := require_login(request):
-        return redirect
-    await excel_file.read()
-    return RedirectResponse("/inventory", status_code=303)
-
-
-@router.get("/license", response_class=HTMLResponse)
-def license_page(request: Request) -> HTMLResponse:
-    """Render the license inventory page."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "licenses": [],
-        "columns": [],
-        "column_widths": {},
-        "lookups": {},
-        "offset": 0,
-        "page": 1,
-        "total_pages": 1,
-        "q": "",
-        "per_page": 25,
-        "table_name": "license",
-        "filters": [],
-        "count": 0,
-    }
-    return templates.TemplateResponse(request, "lisans.html", context)
-
-
-@router.post("/license/add")
-async def license_add(request: Request):
-    """Create or update a software license record."""
-    if redirect := require_login(request):
-        return redirect
-    form = await request.form()
-    db = SessionLocal()
-    try:
-        license_id = form.get("license_id")
-        if license_id:
-            item = db.query(LicenseInventory).get(int(license_id))
-            if item:
-                for field in [
-                    "departman",
-                    "kullanici",
-                    "yazilim_adi",
-                    "lisans_anahtari",
-                    "mail_adresi",
-                    "envanter_no",
-                    "ifs_no",
-                    "tarih",
-                    "islem_yapan",
-                    "notlar",
-                ]:
-                    if field in form:
-                        value = form.get(field)
-                        if field == "tarih":
-                            value = date.fromisoformat(value) if value else None
-                        setattr(item, field, value)
-        else:
-            item = LicenseInventory(
-                departman=form.get("departman"),
-                kullanici=form.get("kullanici"),
-                yazilim_adi=form.get("yazilim_adi"),
-                lisans_anahtari=form.get("lisans_anahtari"),
-                mail_adresi=form.get("mail_adresi"),
-                envanter_no=form.get("envanter_no"),
-                ifs_no=form.get("ifs_no"),
-                tarih=
-                    date.fromisoformat(form.get("tarih"))
-                    if form.get("tarih")
-                    else None,
-                islem_yapan=form.get("islem_yapan"),
-                notlar=form.get("notlar"),
-            )
-            db.add(item)
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/license", status_code=303)
-
-
-@router.post("/license/upload")
-async def license_upload(request: Request, excel_file: UploadFile = File(...)):
-    """Accept a license inventory Excel upload (currently discarded)."""
-    if redirect := require_login(request):
-        return redirect
-    await excel_file.read()
-    return RedirectResponse("/license", status_code=303)
-
-
-@router.get("/accessories", response_class=HTMLResponse)
-def accessories_page(request: Request) -> HTMLResponse:
-    """Render the accessories tracking page."""
-    if redirect := require_login(request):
-        return redirect
-    return templates.TemplateResponse(
-        request, "aksesuar.html", {"items": []}
-    )
-
-
-@router.get("/requests", response_class=HTMLResponse)
-def requests_page(request: Request) -> HTMLResponse:
-    """Render the requests tracking page."""
-    if redirect := require_login(request):
-        return redirect
-    lookups = {
-        "donanim_tipi": [],
-        "marka": [],
-        "model": [],
-        "yazilim_adi": [],
-        "urun_adi": [],
-    }
-    context = {
-        "groups": {},
-        "lookups": lookups,
-        "today": date.today().isoformat(),
-    }
-    return templates.TemplateResponse(request, "talep.html", context)
-
-
-@router.post("/requests/add")
-async def requests_add(request: Request):
-    """Add a request record."""
-    if redirect := require_login(request):
-        return redirect
-    form = await request.form()
-    db = SessionLocal()
-    try:
-        item = RequestItem(
-            kategori=form.get("kategori"),
-            donanim_tipi=form.get("donanim_tipi"),
-            marka=form.get("marka"),
-            model=form.get("model"),
-            yazilim_adi=form.get("yazilim_adi"),
-            urun_adi=form.get("urun_adi") or form.get("model") or "",
-            adet=int(form.get("adet") or 0),
-            tarih=
-                date.fromisoformat(form.get("tarih"))
-                if form.get("tarih")
-                else None,
-            ifs_no=form.get("ifs_no"),
-            aciklama=form.get("aciklama"),
-            talep_acan=request.session.get("user", ""),
-        )
-        db.add(item)
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/requests", status_code=303)
-
-
-@router.get("/profile", response_class=HTMLResponse)
-def profile_page(request: Request) -> HTMLResponse:
-    """Render a simple profile page."""
-    if redirect := require_login(request):
-        return redirect
-    user = {"first_name": "", "last_name": "", "email": ""}
-    return templates.TemplateResponse(request, "profile.html", {"user": user})
-
-
-@router.get("/change-password", response_class=HTMLResponse)
-def change_password_page(request: Request) -> HTMLResponse:
-    """Render change password page."""
-    if redirect := require_login(request):
-        return redirect
-    return templates.TemplateResponse(request, "change_password.html")
-
-
-@router.get("/lists", response_class=HTMLResponse)
-def lists_page(request: Request) -> HTMLResponse:
-    """Render the lists management page."""
-    if redirect := require_login(request):
-        return redirect
-    context = {
-        "brands": [],
-        "locations": [],
-        "types": [],
-        "softwares": [],
-        "factories": [],
-        "departments": [],
-        "blocks": [],
-        "models": [],
-        "printer_brands": [],
-        "printer_models": [],
-        "products": [],
-    }
-    return templates.TemplateResponse(request, "listeler.html", context)
-
-
-@router.post("/lists/add")
-async def lists_add(request: Request, item_type: str = Form(...), name: str = Form(...)):
-    """Add a lookup list item."""
-
-    if redirect := require_login(request):
-        return redirect
-    db = SessionLocal()
-    try:
-        db.add(LookupItem(type=item_type, name=name))
-        db.commit()
-    finally:
-        db.close()
-    return RedirectResponse("/lists", status_code=303)
-
-
-@router.get("/ping")
-def ping(request: Request):
-    """Simple authenticated health check."""
-    if redirect := require_login(request):
-        return redirect
-    return {"status": "ok"}
-
-
-@router.get("/table-columns")
-def table_columns(request: Request, table_name: str):
-    """Return available columns for the requested table."""
-    if redirect := require_login(request):
-        return redirect
-    model = MODEL_MAP.get(table_name)
-    if not model:
-        return {"columns": []}
-    cols = get_table_columns(model.__tablename__)
-    return {"columns": cols}
-
-
-@router.get("/column-settings")
-def column_settings(request: Request, table_name: str):
-    """Fetch stored column settings for a table."""
-    if redirect := require_login(request):
-        return redirect
-    settings = load_settings()
-    return settings.get(table_name, {})
-
-
-@router.post("/column-settings")
-def save_column_settings(request: Request, table_name: str, data: dict):
-    """Persist column settings for a table."""
-    if redirect := require_login(request):
-        return redirect
-    settings = load_settings()
-    settings[table_name] = data
-    save_settings(settings)
-    return {"status": "ok"}
-
-
-@router.get("/column-values")
-def column_values(request: Request, table_name: str, column: str | None = None):
-    """Return distinct values for a column to power client-side filters."""
-    if redirect := require_login(request):
-        return redirect
-    if not column:
-        return {"values": []}
-    model = MODEL_MAP.get(table_name)
-    if not model or not hasattr(model, column):
-        return {"values": []}
-    db = SessionLocal()
-    try:
-        values = [row[0] for row in db.query(getattr(model, column)).distinct()]
-    finally:
-        db.close()
-    return {"values": values}
